feat(food-delivery): code-splitting using webpack
affects: @practice/food-delivery

ISSUES CLOSED: #52
