+++ package.json
@@ -2,6 +2,8 @@
   "license": "BSD 3-Clause",
   "private": true,
   "scripts": {
+    "build": "webpack",
+    "compile": "tsc",
     "lint": "./node_modules/.bin/eslint --fix './**/*.{ts,js}'"
   },
   "devDependencies": {
@@ -17,7 +19,12 @@
     "lerna": "3.22.1",
     "lint-staged": "^10.5.3",
     "prettier": "^2.2.1",
-    "typescript": "^4.1.3"
+    "typescript": "^4.1.3",
+    "eslint-import-resolver-typescript": "^2.4.0",
+    "ts-loader": "^8.0.17",
+    "ts-node": "^9.1.1",
+    "webpack": "^5.24.2",
+    "webpack-cli": "^4.5.0"
   },
   "husky": {
     "hooks": {
