+++ practice/food-delivery/.eslintrc.json
@@ -2,6 +2,12 @@
   "extends": "../../.eslintrc.json",
   "rules": {
     "no-use-before-define": 0,
-    "no-undef": 0
+    "no-undef": 0,
+    "import/extensions": 0
+  },
+  "settings": {
+  "import/resolver": {
+    "typescript": "./src/**/*"
+    }
   }
 }
