+++ practice/food-delivery/package.json
@@ -2,9 +2,10 @@
   "name": "@practice/food-delivery",
   "version": "0.0.0",
   "license": "BSD 3-Clause",
-  "main": "src/index.ts",
+  "main": "./src/index.ts",
   "scripts": {
-    "build": "tsc",
+    "build": "webpack",
+    "compile": "tsc",
     "lint": "eslint"
   }
 }
