+++ practice/food-delivery/public/index.html
@@ -86,7 +86,7 @@
                 </div>
             </div>
             <div class="offer__action">
-                <button modal-data class="btn btn_dark">Связаться с нами</button>
+                <button data-modal class="btn btn_dark">Связаться с нами</button>
             </div>
         </div>
         <div class="container">
@@ -298,6 +298,6 @@
             </div>
         </div>
     </div>
-    <script src='../dist/index.js'></script>
+    <script src='../dist/bundle.js'></script>
 </body>
 </html>
